pyproject.toml
[project]
name = "roboneuron"
version = "0.1.0"
description = "ROS2-native middleware for perception, VLA inference, control, and MCP exposure in embodied AI systems"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ikpy>=3.4.2",
    "json-numpy>=2.1.1",
    "jinja2>=3.1.0",
    "mcp>=1.22.0",
    "numpy<2",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "pydantic>=2.11.0,<3",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
vla = [
    "torch>=2.9.1",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/roboneuron_backends", "src/roboneuron_core", "src/roboneuron_edge"]

[project.scripts]
roboneuron-mcp-perception = "roboneuron_core.cli.mcp_entrypoints:mcp_perception"
roboneuron-mcp-vla = "roboneuron_core.cli.mcp_entrypoints:mcp_vla"
roboneuron-mcp-control = "roboneuron_core.cli.mcp_entrypoints:mcp_control"
roboneuron-mcp-twist = "roboneuron_core.cli.mcp_entrypoints:mcp_twist"
roboneuron-mcp-eef-delta = "roboneuron_core.cli.mcp_entrypoints:mcp_eef_delta"
roboneuron-validate-local = "roboneuron_core.cli.local_validation:main"

[dependency-groups]
dev = [
    "black>=26.0.0",
    "hypothesis>=6.148.0",
    "mypy>=1.18.0",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "setuptools==57.5.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
markers = [
    "integration: integration tests",
    "e2e: end-to-end tests",
    "property: property-based tests",
    "ros: requires ROS runtime",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
files = [
    "src/roboneuron_backends",
    "src/roboneuron_core/adapters",
    "src/roboneuron_core/cli",
    "src/roboneuron_core/kernel",
    "src/roboneuron_core/runtime",
    "src/roboneuron_core/utils",
    "src/roboneuron_edge",
    "tests",
]
exclude = [
    "^third_party/vla_src/",
    "^src/roboneuron_core/servers/",
    "^src/roboneuron_edge/servers/",
]

[[tool.mypy.overrides]]
module = ["roboneuron_core.servers.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["roboneuron_core.adapters.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["roboneuron_edge.servers.*"]
ignore_errors = true
